vault location muster list bug fixed
