+++ src/resources/vault-locations/VaultLocationList.tsx
@@ -1,11 +1,17 @@
 import React, { useMemo, useState } from 'react'
-import { Datagrid, FunctionField, type Identifier, List } from 'react-admin'
+import {
+  Datagrid,
+  FunctionField,
+  type Identifier,
+  List,
+  useListContext
+} from 'react-admin'
 import FlexBox from '../../components/FlexBox'
 import VaultLocationReport from '../../components/VaultLocationReport'
 import useCanAccess from '../../hooks/useCanAccess'
 import * as constants from '../../constants'
-import { IconButton } from '@mui/material'
-import { History } from '@mui/icons-material'
+import { Button, IconButton } from '@mui/material'
+import { Article, History } from '@mui/icons-material'
 import ResourceHistoryModal from '../../components/ResourceHistory'
 import useDoubleClick from '../../hooks/useDoubleClick'
 
@@ -20,7 +26,9 @@ export default function VaultLocationList(): React.ReactElement {
 const DataList = (): React.ReactElement => {
   const { hasAccess } = useCanAccess()
   const [open, setOpen] = useState<boolean>()
+  const [openMusterList, setOpenMusterList] = useState(false)
   const [record, setRecord] = useState<ReferenceItem>()
+  const { selectedIds } = useListContext()
   const handleRowClick = useDoubleClick(constants.R_VAULT_LOCATION)
 
   const hasWriteAccess = hasAccess(constants.R_VAULT_LOCATION, {
@@ -34,11 +42,21 @@ const DataList = (): React.ReactElement => {
     [record]
   )
 
+  const handleOpen = (open: boolean) => () => {
+    setOpenMusterList(open)
+  }
+
   const BulkActions = (): React.ReactElement => {
     return (
       <>
         <FlexBox>
-          <VaultLocationReport />
+          <Button
+            startIcon={<Article />}
+            sx={{ lineHeight: '1.5' }}
+            size='small'
+            onClick={handleOpen(true)}>
+            Location Muster List
+          </Button>
         </FlexBox>
       </>
     )
@@ -79,6 +97,11 @@ const DataList = (): React.ReactElement => {
         close={() => {
           setOpen(false)
         }}
+      />
+      <VaultLocationReport
+        open={openMusterList}
+        handleOpen={handleOpen}
+        selectedIds={selectedIds}
       />
     </>
   )
